+++ wutengine_ecs/src/world/queries.rs
@@ -4,7 +4,7 @@ use core::cell::UnsafeCell;
 use wutengine_core::{EntityId, ReadWriteDescriptor};
 use wutengine_util_macro::make_combined_query_tuples;
 
-use crate::archetype::ArchetypeId;
+use crate::archetype::{self, ArchetypeId};
 use crate::vec::AnyVec;
 
 use super::World;
@@ -24,11 +24,13 @@ impl World {
         C: CombinedQuery<'a>,
         F: Fn(EntityId, C) -> O,
     {
-        let type_ids = C::get_type_ids();
+        let descriptors = C::get_query_descriptors();
 
-        assert_unique_type_ids(&type_ids);
+        assert_unique_type_ids(&descriptors);
 
-        let archetype_ids = self.archetype_ids_for(&type_ids);
+        let archetype_ids = self.archetype_ids_for(&descriptors);
+
+        let type_ids: Vec<TypeId> = descriptors.iter().map(|desc| desc.type_id).collect();
 
         let mut output = Vec::new();
 
@@ -47,32 +49,46 @@ impl World {
         output
     }
 
-    fn archetype_ids_for(&self, type_ids: &[TypeId]) -> Vec<ArchetypeId> {
-        let mut init_archetypes = match self.type_containers.get(&type_ids[0]) {
-            Some(archetypes) => archetypes.clone(),
-            None => return Vec::new(),
-        };
+    fn archetype_ids_for(&self, queried_components: &[QueryDescriptor]) -> Vec<ArchetypeId> {
+        let required = queried_components
+            .iter()
+            .filter(|qc| matches!(qc.query_type, QueryType::Required));
 
-        for type_id in &type_ids[1..] {
-            let containing_archetypes = match self.type_containers.get(type_id) {
-                Some(archetypes) => archetypes,
+        // Get only the archetype IDs for the required components.
+        // If these archetypes happen to have one or more of the optional components,
+        // then hey, we're in luck!
+        self.archetype_ids_for_required(required)
+    }
+
+    fn archetype_ids_for_required<'a>(
+        &self,
+        queried_components: impl Iterator<Item = &'a QueryDescriptor>,
+    ) -> Vec<ArchetypeId> {
+        let mut archetype_set: Option<Vec<ArchetypeId>> = None;
+
+        for queried_component in queried_components {
+            let component_archetypes = match self.type_containers.get(&queried_component.type_id) {
+                Some(archetype) => archetype,
                 None => return Vec::new(),
             };
 
-            init_archetypes.retain(|e| containing_archetypes.contains(e));
+            if let Some(archetype_set) = &mut archetype_set {
+                archetype_set.retain(|e| component_archetypes.contains(e));
+            } else {
+                archetype_set = Some(component_archetypes.clone());
+            }
 
-            if init_archetypes.is_empty() {
-                // Short-circuit
+            if archetype_set.as_ref().unwrap().is_empty() {
                 return Vec::new();
             }
         }
 
-        init_archetypes
+        archetype_set.unwrap_or_default()
     }
 }
 
 #[track_caller]
-pub fn assert_unique_type_ids(ids: &[TypeId]) {
+pub fn assert_unique_type_ids(ids: &[QueryDescriptor]) {
     let duplicate = get_first_duplicate_type_id(ids);
 
     if let Some(duplicate) = duplicate {
@@ -80,13 +96,13 @@ pub fn assert_unique_type_ids(ids: &[TypeId]) {
     }
 }
 
-pub fn get_first_duplicate_type_id(ids: &[TypeId]) -> Option<TypeId> {
+pub fn get_first_duplicate_type_id(ids: &[QueryDescriptor]) -> Option<TypeId> {
     for i in 0..ids.len() {
         for j in (i + 1)..ids.len() {
             debug_assert_ne!(i, j);
 
-            if ids[i] == ids[j] {
-                return Some(ids[i]);
+            if ids[i].type_id == ids[j].type_id {
+                return Some(ids[i].type_id);
             }
         }
     }
@@ -97,8 +113,9 @@ pub fn get_first_duplicate_type_id(ids: &[TypeId]) -> Option<TypeId> {
 pub trait Queryable<'q>: Sized {
     type Inner: Any;
     const READ_ONLY: bool;
+    const QUERY_TYPE: QueryType;
 
-    fn from_anyvec<'a: 'q>(cell: &'a UnsafeCell<AnyVec>) -> Vec<Self>;
+    fn from_anyvec<'a: 'q>(num_entities: usize, cell: Option<&'a UnsafeCell<AnyVec>>) -> Vec<Self>;
 }
 
 impl<'q, T> Queryable<'q> for &'q T
@@ -107,13 +124,21 @@ where
 {
     type Inner = T;
     const READ_ONLY: bool = true;
+    const QUERY_TYPE: QueryType = QueryType::Required;
 
-    fn from_anyvec<'a: 'q>(cell: &'a UnsafeCell<AnyVec>) -> Vec<Self> {
+    fn from_anyvec<'a: 'q>(num_entities: usize, cell: Option<&'a UnsafeCell<AnyVec>>) -> Vec<Self> {
         let cell_ref = unsafe {
-            cell.get()
+            cell.expect("None AnyVec given to required non-mutable queryable")
+                .get()
                 .as_ref::<'q>()
                 .expect("UnsafeCell returned nullptr")
         };
+
+        debug_assert_eq!(
+            num_entities,
+            cell_ref.len(),
+            "Unexpected amount of components in AnyVec"
+        );
 
         let mut output = Vec::with_capacity(cell_ref.len());
 
@@ -131,13 +156,21 @@ where
 {
     type Inner = T;
     const READ_ONLY: bool = false;
+    const QUERY_TYPE: QueryType = QueryType::Required;
 
-    fn from_anyvec<'a: 'q>(cell: &'a UnsafeCell<AnyVec>) -> Vec<Self> {
+    fn from_anyvec<'a: 'q>(num_entities: usize, cell: Option<&'a UnsafeCell<AnyVec>>) -> Vec<Self> {
         let cell_ref = unsafe {
-            cell.get()
+            cell.expect("None AnyVec given to required mutable queryable")
+                .get()
                 .as_mut::<'q>()
                 .expect("UnsafeCell returned nullptr")
         };
+
+        debug_assert_eq!(
+            num_entities,
+            cell_ref.len(),
+            "Unexpected amount of components in AnyVec"
+        );
 
         let mut output = Vec::with_capacity(cell_ref.len());
 
@@ -149,27 +182,64 @@ where
     }
 }
 
+impl<'q, T> Queryable<'q> for Option<T>
+where
+    T: Queryable<'q>,
+{
+    type Inner = T::Inner;
+    const READ_ONLY: bool = T::READ_ONLY;
+    const QUERY_TYPE: QueryType = QueryType::Optional;
+
+    fn from_anyvec<'a: 'q>(num_entities: usize, cell: Option<&'a UnsafeCell<AnyVec>>) -> Vec<Self> {
+        match cell {
+            Some(cell) => {
+                let found_vec = T::from_anyvec(num_entities, Some(cell));
+
+                found_vec.into_iter().map(Some).collect()
+            }
+            None => Vec::from_iter(std::iter::repeat_with(|| None).take(num_entities)),
+        }
+    }
+}
+
 pub trait CombinedQuery<'q>: Sized {
-    fn get_type_ids() -> Vec<TypeId>;
-    fn get_descriptors() -> Vec<ReadWriteDescriptor>;
+    fn get_query_descriptors() -> Vec<QueryDescriptor>;
+    fn get_read_write_descriptors() -> Vec<ReadWriteDescriptor>;
     fn do_callback<F, O>(
         entities: &[EntityId],
-        cells: Vec<&'q UnsafeCell<AnyVec>>,
+        cells: Vec<Option<&'q UnsafeCell<AnyVec>>>,
         callback: F,
     ) -> Vec<O>
     where
         F: Fn(EntityId, Self) -> O;
 }
 
+#[derive(Debug, Clone, Copy)]
+pub struct QueryDescriptor {
+    pub type_id: TypeId,
+    pub query_type: QueryType,
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum QueryType {
+    Required,
+    Optional,
+    //TODO: NOT filters
+    // Not,
+}
+
 impl<'q, T> CombinedQuery<'q> for T
 where
     T: Queryable<'q>,
 {
-    fn get_type_ids() -> Vec<TypeId> {
-        vec![TypeId::of::<T::Inner>()]
+    fn get_query_descriptors() -> Vec<QueryDescriptor> {
+        vec![QueryDescriptor {
+            type_id: TypeId::of::<T::Inner>(),
+            query_type: T::QUERY_TYPE,
+        }]
     }
 
-    fn get_descriptors() -> Vec<ReadWriteDescriptor> {
+    fn get_read_write_descriptors() -> Vec<ReadWriteDescriptor> {
         vec![ReadWriteDescriptor {
             type_id: TypeId::of::<T::Inner>(),
             read_only: T::READ_ONLY,
@@ -177,17 +247,17 @@ where
     }
 
     fn do_callback<F, O>(
-        entites: &[EntityId],
-        cells: Vec<&'q UnsafeCell<AnyVec>>,
+        entities: &[EntityId],
+        cells: Vec<Option<&'q UnsafeCell<AnyVec>>>,
         callback: F,
     ) -> Vec<O>
     where
         F: Fn(EntityId, Self) -> O,
     {
-        let refs = T::from_anyvec(cells[0]);
+        let refs = T::from_anyvec(entities.len(), cells[0]);
         let mut outputs = Vec::with_capacity(refs.len());
 
-        for (args, &entity) in refs.into_iter().zip(entites) {
+        for (args, &entity) in refs.into_iter().zip(entities) {
             outputs.push(callback(entity, args));
         }