+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 This file documents the most important changes for each released version.
 
 ## [v0.0.7]
+- Initial support for optional components in ECS queries
 
 ## [v0.0.6]
 - Complete refactor of core ECS system. We now use archetypes. It was a lot of work.